+++ halo2_proofs/src/lib.rs
@@ -20,4 +20,4 @@ pub mod dev;
 mod helpers;
 
 /// | # of bytes dedicated to vk footers
-pub const COSMWASM_FOOTER_LENGTH: usize = 46;
+pub const COSMWASM_FOOTER_LENGTH: usize = 80;